requirements.txt
# Core dependencies
scapy==2.5.0
python-nmap==0.7.1
dnspython==2.6.1
requests==2.31.0
aiohttp==3.9.5
colorama==0.4.6
rich>=13.7.0
pyyaml==6.0.1
psutil==5.9.8
click>=8.1.7

# Security
cryptography>=42.0.2
pyOpenSSL==24.0.0
certifi==2024.2.2

# Network utilities
netifaces==0.11.0
ipaddress==1.0.23
python-whois==0.8.0

# Rich extensions
rich-cli==1.8.0
rich-rst==1.1.7
rich-sparkline==0.1.0

# Report generation
openpyxl>=3.1.2
reportlab>=4.1.0
pandas>=2.2.0
matplotlib==3.8.3
seaborn==0.13.2